Move ledger genesis to application
